Implement CommandType enum
Issue #65 - https://github.com/cs2103aug2014-f11-3j/main/issues/65

Why is this change necessary?
For a general Command object, a field is needed to identify the precise
command type. For example, "add something" is a command and can be
represented as a Command object. However, to execute this add command,
one needs to know that it is an add command, i.e. we somehow need a way
to tell the execution method that the Command object is an add command,
as opposed to a delete command.

How does it address the issue?
A command type can be represented by a CommandType enum in the Command
object. One just has to access the commandType field of the Command
object. For those who wonder why a new class for every type of command
is not created, the reason is that the command log in the Database class
does not differentiate among command types. Furthermore, many commands
are associated with tasks, thus it may be useful to have a general
Command class storing with associated tasks. Polymorphism is possible,
but having a field to identify the command type is by far the simplest
solution.

What side effects does this change have?
- UserInterface class will need to fill in the commandType field in a
  Command object when parsing user commands.
- Logic class will identify the command type through the commandType
  field and execute it appropriately.