Implement task-to-string conversion method
https://github.com/cs2103aug2014-f11-3j/main/issues/70

Why is this change necessary?
All tasks (with their associated information) need to be logged into a
text file.

How does it address the issue?
Having a method to convert a Task object and its relevant information to
a string aids writing the latter into a text file.

What side effects does this change have?
GUI may want to use this task-to-string conversion method to display
tasks on-screen.