Implement logging of task ID
https://github.com/cs2103aug2014-f11-3j/main/issues/86

Why is this change necessary?
Users should be able to use task IDs for task manipulation just from
looking at the log file. Hence task IDs need to be logged like any other
task attribute.

How does it address the issue?
Implementing read/write functionalities to/from task log (text file) for
task IDs allows user to view the log file in real time and manipulate
tasks.

What side effects does this change have?
Currently, task IDs are dynamically set to be the index of the arraylist
in which they are stored. So reading task IDs from the task log is
redundant for now. However, task IDs may be fixed uniquely in the future
and it becomes important for TaskBuddy to read in these IDs along with
their tasks.

Close #86.